Add mylib C library to lua modules
This commit introduces a new C-library called 'mylib' which is meant to be used as a module in lua scripts. 'mylib' provides a simple addition function. A test lua script is provided for implementation verification. The Makefile for compiling the shared library with cc on macOS is also added. This allows more complex functions, not possible in lua, to be defined in C and used in lua scripts.